requirements.txt
streamlit
keras
numpy
pillow
tensorflow
